Select org roles using modal, not select-in-table
